# src/main/java/atlassports/api/ResourcesApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.18.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package atlassports.api;

import atlassports.model.dto.ResourceDto;
import atlassports.model.dto.UpsertResourceDto;
import atlassports.model.openapi.ResourcesGet200Response;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import jakarta.validation.constraints.Max;
import jakarta.validation.constraints.Min;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.lang.Nullable;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.context.request.NativeWebRequest;

import javax.annotation.processing.Generated;
import java.util.Optional;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-12-24T16:37:25.216838800+01:00[Europe/Skopje]", comments = "Generator version: 7.18.0")
@Validated
@Tag(name = "Resources", description = "Venue resources")
public interface ResourcesApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_RESOURCES_GET = "/resources";

    /**
     * GET /resources : Get all resources
     *
     * @param page (optional, default to 0)
     * @param size (optional, default to 20)
     * @param sort (optional)
     * @return Paginated list of past bookings (status code 200)
     */
    @Operation(
            operationId = "resourcesGet",
            summary = "Get all resources",
            tags = {"Resources"},
            responses = {
                    @ApiResponse(responseCode = "200", description = "Paginated list of past bookings", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = ResourcesGet200Response.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.GET,
            value = ResourcesApi.PATH_RESOURCES_GET,
            produces = {"application/json"}
    )
    default ResponseEntity<Page<ResourceDto>> resourcesGet(
            Pageable page
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType : MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"size\" : 6, \"last\" : true, \"totalPages\" : 5, \"page\" : 0, \"content\" : [ null, null ], \"first\" : true, \"totalElements\" : 1 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_RESOURCES_POST = "/resources";

    /**
     * POST /resources : Create resource
     *
     * @param body (required)
     * @return Resource created (status code 201)
     */
    @Operation(
            operationId = "resourcesPost",
            summary = "Create resource",
            tags = {"Resources"},
            responses = {
                    @ApiResponse(responseCode = "201", description = "Resource created", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.POST,
            value = ResourcesApi.PATH_RESOURCES_POST,
            produces = {"application/json"},
            consumes = {"application/json"}
    )
    default ResponseEntity<ResourceDto> resourcesPost(
            @Valid @RequestBody UpsertResourceDto upsertResourceDto
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}