# src/main/java/atlassports/api/StatisticsApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.18.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package atlassports.api;

import atlassports.model.dto.SystemStatisticsDto;
import atlassports.model.dto.TenantStatisticsDto;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.constraints.NotNull;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.context.request.NativeWebRequest;

import javax.annotation.processing.Generated;
import java.util.Optional;
import java.util.UUID;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-12-24T16:37:25.216838800+01:00[Europe/Skopje]", comments = "Generator version: 7.18.0")
@Validated
@Tag(name = "Statistics", description = "Reporting and analytics")
public interface StatisticsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_STATISTICS_SYSTEM_GET = "/statistics/system";

    /**
     * GET /statistics/system : Get system-wide statistics (Admin only)
     *
     * @return System statistics (status code 200)
     */
    @Operation(
            operationId = "statisticsSystemGet",
            summary = "Get system-wide statistics (Admin only)",
            tags = {"Statistics"},
            responses = {
                    @ApiResponse(responseCode = "200", description = "System statistics", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.GET,
            value = StatisticsApi.PATH_STATISTICS_SYSTEM_GET,
            produces = {"application/json"}
    )
    default ResponseEntity<SystemStatisticsDto> statisticsSystemGet(

    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_STATISTICS_TENANTS_TENANT_ID_GET = "/statistics/tenants/{tenantId}";

    /**
     * GET /statistics/tenants/{tenantId} : Get statistics for a tenant
     *
     * @param tenantId (required)
     * @return Tenant statistics (status code 200)
     */
    @Operation(
            operationId = "statisticsTenantsTenantIdGet",
            summary = "Get statistics for a tenant",
            tags = {"Statistics"},
            responses = {
                    @ApiResponse(responseCode = "200", description = "Tenant statistics", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.GET,
            value = StatisticsApi.PATH_STATISTICS_TENANTS_TENANT_ID_GET,
            produces = {"application/json"}
    )
    default ResponseEntity<TenantStatisticsDto> statisticsTenantsTenantIdGet(
            @NotNull @Parameter(name = "tenantId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("tenantId") Long tenantId
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}