# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Atlas Sports API
  version: 1.0.0
servers:
- url: http://localhost:8080/api
...
paths:
...
  /tenants/{tenantId}:
    get:
      parameters:
      - explode: false
        in: path
        name: tenantId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantDto"
          description: Tenant details
      summary: Get tenant details
      tags:
      - Tenants
      x-accepts:
      - application/json
      x-tags:
      - tag: Tenants
    patch:
      parameters:
      - explode: false
        in: path
        name: tenantId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertTenantDto"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantDto"
          description: Updated tenant
      summary: Update tenant
      tags:
      - Tenants
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Tenants
...
  /resources:
    get:
      parameters:
      - explode: true
        in: query
        name: page
        required: false
        schema:
          default: 0
          minimum: 0
          type: integer
        style: form
      - explode: true
        in: query
        name: size
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: integer
        style: form
      - explode: true
        in: query
        name: sort
        required: false
        schema:
          example: "createdAt,desc"
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_resources_get_200_response"
          description: Paginated list of past bookings
      summary: Get all resources
      tags:
      - Resources
      x-accepts:
      - application/json
      x-tags:
      - tag: Resources
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertResourceDto"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceDto"
          description: Resource created
      summary: Create resource
      tags:
      - Resources
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Resources
...
  /bookings/tenants/{tenantId}:
    get:
      parameters:
      - explode: false
        in: path
        name: tenantId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/BookingDto"
                type: array
          description: List of bookings
      summary: Get all bookings for tenant
      tags:
      - Bookings
      x-accepts:
      - application/json
      x-tags:
      - tag: Bookings
  /bookings/me/upcoming:
    get:
      parameters:
      - explode: true
        in: query
        name: page
        required: false
        schema:
          default: 0
          minimum: 0
          type: integer
        style: form
      - explode: true
        in: query
        name: size
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: integer
        style: form
      - explode: true
        in: query
        name: sort
        required: false
        schema:
          example: "createdAt,desc"
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_bookings_me_upcoming_get_200_response"
          description: List of upcoming bookings
      summary: Get upcoming bookings for current user
      tags:
      - Bookings
      x-accepts:
      - application/json
      x-tags:
      - tag: Bookings
  /bookings/me/history:
    get:
      parameters:
      - explode: true
        in: query
        name: page
        required: false
        schema:
          default: 0
          minimum: 0
          type: integer
        style: form
      - explode: true
        in: query
        name: size
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: integer
        style: form
      - explode: true
        in: query
        name: sort
        required: false
        schema:
          example: "createdAt,desc"
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_bookings_me_upcoming_get_200_response"
          description: Paginated list of past bookings
      summary: Get past bookings for current user
      tags:
      - Bookings
      x-accepts:
      - application/json
      x-tags:
      - tag: Bookings
...
  /statistics/tenants/{tenantId}:
    get:
      parameters:
      - explode: false
        in: path
        name: tenantId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantStatisticsDto"
          description: Tenant statistics
      summary: Get statistics for a tenant
      tags:
      - Statistics
      x-accepts:
      - application/json
      x-tags:
      - tag: Statistics
...
components:
...
  schemas:
...
    TenantDto:
      type: object
    UpsertTenantDto:
      type: object
...
    ResourceDto:
      type: object
    UpsertResourceDto:
      type: object
    BookingDto:
      type: object
...
    TenantStatisticsDto:
      type: object
...
    PageResponse:
      properties:
        content:
          items:
            type: object
          type: array
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
      type: object
...
    _resources_get_200_response:
      allOf:
      - $ref: "#/components/schemas/PageResponse"
      - properties:
          content:
            items:
              $ref: "#/components/schemas/ResourceDto"
            type: array
        type: object
      example:
        size: 6
        last: true
        totalPages: 5
        page: 0
        content:
        - null
        - null
        first: true
        totalElements: 1
    _bookings_me_upcoming_get_200_response:
      allOf:
      - $ref: "#/components/schemas/PageResponse"
      - properties:
          content:
            items:
              $ref: "#/components/schemas/BookingDto"
            type: array
        type: object
      example:
        size: 6
        last: true
        totalPages: 5
        page: 0
        content:
        - null
        - null
        first: true
        totalElements: 1